Merge main m3 (#1278)
* Initial version

Change-Id: I068d920de50bc8ac686392826bf4739803420343

* Adds content padding

Change-Id: I9315ccb38e5b6233bebac1c535c2361dcbbca97c

* Adds license headers

Change-Id: I43a982d53264f72e36dafc08901b46cbceaeac42

* Updates libraries

Change-Id: I86993a3a727f9fdf700a34a84326ee39c376e5f2

* Updates validator

Updates the validator to the latest version of the validator - this version is no longer a noop and performs actual checks.

Ideally this binary will be swapped out for a library version by launch

Change-Id: I4e5e17a611860f020f1dae34caa158c9b0ada541

* Add build workflows for Wear OS modules

This commit adds new GitHub Actions workflows to build the
Wear OS Compose Starter and DataLayer modules. These workflows
automate the build process for the following:

- Compose Starter app
- DataLayer Application app
- DataLayer Wearable app

The workflows are triggered on push, pull request, and
workflow_dispatch events on the main branch. They set up JDK 17,
Gradle, make gradlew executable, and build the specified module.

* Update Gradle version to 8.1.4

    ./gradlew wrapper --gradle-version latest
    ./gradlew wrapper

This commit updates the Gradle version used by the
ComposeStarter and DataLayer modules from 8.1.3 to 8.1.4.
This includes changes to:

- `gradle-wrapper.jar` (binary file update)
- `gradle-wrapper.properties` (distribution URL update)
- `gradlew` (classpath and jar execution changes)
- `gradlew.bat` (classpath and jar execution changes)

* Update Gradle version in validator-plugin

This commit updates the Gradle version used by the
`validator-plugin` module from 8.13 to 8.14. This includes
changes to:

- `gradle-wrapper.jar` (binary file update)
- `gradle-wrapper.properties` (distribution URL update)
- `gradlew` (classpath and jar execution changes)
- `gradlew.bat` (classpath and jar execution changes)

* Adds use of Validator CLI via maven

Change-Id: If5bc5a04587a3788a8a1150c77dd2d1fbbe326e4

* Adds README.md and updates libraries

Change-Id: Ic2410bb00c0cbc0ae548a568237fe5a372e2d3cc

* Minor gradle fix

Fixes an issue where individual variant tasks such as assembleDebug and assembleRelease work but specifying the overall task assemble will not because the output path used by one of the tasks was not specific enough so created an unintended dependency between the variants.

Change-Id: Ife0835f7800f62301549bb8f7e28539bec0142a9

* Adds Photos WFF samples

* Fix Gradle issues

* Fix errors

Change-Id: Ia2cb690401b9c678ee733b7312f0355ed0f886f7

* Update all dependencies (#1268)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Update all dependencies (#1275)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

---------

Co-authored-by: Automerger <automerger@devrel-build.iam.gserviceaccount.com>
Co-authored-by: garan <garan@google.com>
Co-authored-by: cartland <c.cartland@gmail.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>